Separate Polynomial from kzg module
